[Transform] Implement exponential backoff for transform state persistence retrying
