+++ x-pack/plugin/transform/src/test/java/org/elasticsearch/xpack/transform/transforms/TransformIndexerFailureOnStatePersistenceTests.java
@@ -281,7 +281,6 @@ public class TransformIndexerFailureOnStatePersistenceTests extends ESTestCase {
                     }
                 );
             }
-
         }
 
         // test reset on success
